Minor editor refactoring
