+++ spyder/plugins/editor/widgets/codeeditor.py
@@ -28,24 +28,19 @@ import re
 import sre_constants
 import sys
 import textwrap
-import time
 
 # Third party imports
 from diff_match_patch import diff_match_patch
 from IPython.core.inputtransformer2 import TransformerManager
 from qtpy.compat import to_qvariant
-from qtpy.QtCore import (QEvent, QPoint, QRegExp, Qt, QTimer, QThread, QUrl,
-                         Signal, Slot)
-from qtpy.QtGui import (QColor, QCursor, QFont, QIntValidator,
-                        QKeySequence, QPaintEvent, QPainter, QMouseEvent,
-                        QTextCharFormat, QTextCursor, QDesktopServices,
+from qtpy.QtCore import (QEvent, QRegExp, Qt, QTimer, QThread, QUrl, Signal,
+                         Slot)
+from qtpy.QtGui import (QColor, QCursor, QFont, QKeySequence, QPaintEvent,
+                        QPainter, QMouseEvent, QTextCursor, QDesktopServices,
                         QKeyEvent, QTextDocument, QTextFormat, QTextOption,
-                        QTextFrameFormat)
-from qtpy.QtPrintSupport import QPrinter
-from qtpy.QtWidgets import (QApplication, QDialog, QDialogButtonBox,
-                            QGridLayout, QHBoxLayout, QLabel,
-                            QLineEdit, QMenu, QMessageBox, QSplitter,
-                            QToolTip, QVBoxLayout, QScrollBar)
+                        QTextCharFormat)
+from qtpy.QtWidgets import (QApplication, QMenu, QMessageBox, QSplitter,
+                            QToolTip, QScrollBar)
 from spyder_kernels.utils.dochelpers import getobj
 from three_merge import merge
 
@@ -73,28 +68,27 @@ from spyder.plugins.editor.panels import (ClassFunctionDropdown,
                                           FoldingPanel, IndentationGuide,
                                           LineNumberArea, PanelsManager,
                                           ScrollFlagArea)
-from spyder.plugins.editor.utils.editor import (TextHelper, BlockUserData)
+from spyder.plugins.editor.utils.editor import (TextHelper, BlockUserData,
+                                                get_file_language)
 from spyder.plugins.editor.utils.debugger import DebuggerManager
-# from spyder.plugins.editor.utils.folding import IndentFoldDetector, FoldScope
 from spyder.plugins.editor.utils.kill_ring import QtKillRing
 from spyder.plugins.editor.utils.languages import ALL_LANGUAGES, CELL_LANGUAGES
 from spyder.plugins.editor.panels.utils import (
     merge_folding, collect_folding_regions)
 from spyder.plugins.completion.manager.decorators import (
     request, handles, class_register)
+from spyder.plugins.editor.widgets.codeeditor_widgets import GoToLineDialog
 from spyder.plugins.editor.widgets.base import TextEditBaseWidget
-from spyder.plugins.outlineexplorer.languages import PythonCFM
 from spyder.plugins.outlineexplorer.api import (OutlineExplorerData as OED,
                                                 is_cell_header)
 from spyder.py3compat import PY2, to_text_string, is_string, is_text_string
-from spyder.utils import encoding, programs, sourcecode
+from spyder.utils import encoding, sourcecode
 from spyder.utils import icon_manager as ima
 from spyder.utils import syntaxhighlighters as sh
 from spyder.utils.qthelpers import (add_actions, create_action, file_uri,
                                     mimedata2url, start_file)
 from spyder.utils.vcs import get_git_remotes, remote_to_url
 from spyder.utils.qstringhelpers import qstring_length
-from spyder.widgets.helperwidgets import MessageCheckBox
 
 
 try:
@@ -110,155 +104,30 @@ logger = logging.getLogger(__name__)
 # the up/down arrow keys.
 UPDATE_DECORATIONS_TIMEOUT = 500  # milliseconds
 
+
 # %% This line is for cell execution testing
-def is_letter_or_number(char):
-    """Returns whether the specified unicode character is a letter or a number.
-    """
-    cat = category(char)
-    return cat.startswith('L') or cat.startswith('N')
-
-# =============================================================================
-# Go to line dialog box
-# =============================================================================
-class GoToLineDialog(QDialog):
-    def __init__(self, editor):
-        QDialog.__init__(self, editor, Qt.WindowTitleHint
-                         | Qt.WindowCloseButtonHint)
-
-        # Destroying the C++ object right after closing the dialog box,
-        # otherwise it may be garbage-collected in another QThread
-        # (e.g. the editor's analysis thread in Spyder), thus leading to
-        # a segmentation fault on UNIX or an application crash on Windows
-        self.setAttribute(Qt.WA_DeleteOnClose)
-
-        self.lineno = None
-        self.editor = editor
-
-        self.setWindowTitle(_("Editor"))
-        self.setModal(True)
-
-        label = QLabel(_("Go to line:"))
-        self.lineedit = QLineEdit()
-        validator = QIntValidator(self.lineedit)
-        validator.setRange(1, editor.get_line_count())
-        self.lineedit.setValidator(validator)
-        self.lineedit.textChanged.connect(self.text_has_changed)
-        cl_label = QLabel(_("Current line:"))
-        cl_label_v = QLabel("<b>%d</b>" % editor.get_cursor_line_number())
-        last_label = QLabel(_("Line count:"))
-        last_label_v = QLabel("%d" % editor.get_line_count())
-
-        glayout = QGridLayout()
-        glayout.addWidget(label, 0, 0, Qt.AlignVCenter | Qt.AlignRight)
-        glayout.addWidget(self.lineedit, 0, 1, Qt.AlignVCenter)
-        glayout.addWidget(cl_label, 1, 0, Qt.AlignVCenter | Qt.AlignRight)
-        glayout.addWidget(cl_label_v, 1, 1, Qt.AlignVCenter)
-        glayout.addWidget(last_label, 2, 0, Qt.AlignVCenter | Qt.AlignRight)
-        glayout.addWidget(last_label_v, 2, 1, Qt.AlignVCenter)
-
-        bbox = QDialogButtonBox(QDialogButtonBox.Ok|QDialogButtonBox.Cancel,
-                                Qt.Vertical, self)
-        bbox.accepted.connect(self.accept)
-        bbox.rejected.connect(self.reject)
-        btnlayout = QVBoxLayout()
-        btnlayout.addWidget(bbox)
-        btnlayout.addStretch(1)
-
-        ok_button = bbox.button(QDialogButtonBox.Ok)
-        ok_button.setEnabled(False)
-        self.lineedit.textChanged.connect(
-                     lambda text: ok_button.setEnabled(len(text) > 0))
-
-        layout = QHBoxLayout()
-        layout.addLayout(glayout)
-        layout.addLayout(btnlayout)
-        self.setLayout(layout)
-
-        self.lineedit.setFocus()
-
-    def text_has_changed(self, text):
-        """Line edit's text has changed"""
-        text = to_text_string(text)
-        if text:
-            self.lineno = int(text)
-        else:
-            self.lineno = None
-
-    def get_line_number(self):
-        """Return line number"""
-        # It is import to avoid accessing Qt C++ object as it has probably
-        # already been destroyed, due to the Qt.WA_DeleteOnClose attribute
-        return self.lineno
-
-
-#===============================================================================
-# CodeEditor widget
-#===============================================================================
-def get_file_language(filename, text=None):
-    """Get file language from filename"""
-    ext = osp.splitext(filename)[1]
-    if ext.startswith('.'):
-        ext = ext[1:] # file extension with leading dot
-    language = ext
-    if not ext:
-        if text is None:
-            text, _enc = encoding.read(filename)
-        for line in text.splitlines():
-            if not line.strip():
-                continue
-            if line.startswith('#!'):
-               shebang = line[2:]
-               if 'python' in shebang:
-                   language = 'python'
-            else:
-                break
-    return language
-
-
-def count_leading_empty_lines(cell):
-    """Count the number of leading empty cells."""
-    lines = cell.splitlines(keepends=True)
-    if not lines:
-        return 0
-    for i, line in enumerate(lines):
-        if line and not line.isspace():
-            return i
-    return len(lines)
-
-
-def ipython_to_python(code):
-    """Transform IPython code to python code."""
-    tm = TransformerManager()
-    number_empty_lines = count_leading_empty_lines(code)
-    try:
-        code = tm.transform_cell(code)
-    except SyntaxError:
-        return code
-    return '\n' * number_empty_lines + code
-
-
 @class_register
 class CodeEditor(TextEditBaseWidget):
     """Source Code Editor Widget based exclusively on Qt"""
 
     LANGUAGES = {
-        'Python': (sh.PythonSH, '#', PythonCFM),
-        'IPython': (sh.IPythonSH, '#', PythonCFM),
-        'Cython': (sh.CythonSH, '#', PythonCFM),
-        'Fortran77': (sh.Fortran77SH, 'c', None),
-        'Fortran': (sh.FortranSH, '!', None),
-        'Idl': (sh.IdlSH, ';', None),
-        'Diff': (sh.DiffSH, '', None),
-        'GetText': (sh.GetTextSH, '#', None),
-        'Nsis': (sh.NsisSH, '#', None),
-        'Html': (sh.HtmlSH, '', None),
-        'Yaml': (sh.YamlSH, '#', None),
-        'Cpp': (sh.CppSH, '//', None),
-        'OpenCL': (sh.OpenCLSH, '//', None),
-        'Enaml': (sh.EnamlSH, '#', PythonCFM),
-        'Markdown': (sh.MarkdownSH, '#', None),
+        'Python': (sh.PythonSH, '#'),
+        'IPython': (sh.IPythonSH, '#'),
+        'Cython': (sh.CythonSH, '#'),
+        'Fortran77': (sh.Fortran77SH, 'c'),
+        'Fortran': (sh.FortranSH, '!'),
+        'Idl': (sh.IdlSH, ';'),
+        'Diff': (sh.DiffSH, ''),
+        'GetText': (sh.GetTextSH, '#'),
+        'Nsis': (sh.NsisSH, '#'),
+        'Html': (sh.HtmlSH, ''),
+        'Yaml': (sh.YamlSH, '#'),
+        'Cpp': (sh.CppSH, '//'),
+        'OpenCL': (sh.OpenCLSH, '//'),
+        'Enaml': (sh.EnamlSH, '#'),
+        'Markdown': (sh.MarkdownSH, '#'),
         # Every other language
-        'None': (sh.TextSH, '', None),
+        'None': (sh.TextSH, ''),
     }
 
     TAB_ALWAYS_INDENTS = (
@@ -278,7 +147,6 @@ class CodeEditor(TextEditBaseWidget):
     sig_breakpoints_saved = Signal()
     sig_filename_changed = Signal(str)
     sig_bookmarks_changed = Signal()
-    get_completions = Signal(bool)
     go_to_definition = Signal(str, int, int)
     sig_show_object_info = Signal(int)
     sig_run_selection = Signal()
@@ -286,7 +154,6 @@ class CodeEditor(TextEditBaseWidget):
     sig_run_cell = Signal()
     sig_re_run_last_cell = Signal()
     sig_debug_cell = Signal()
-    go_to_definition_regex = Signal(str, int, int)
     sig_cursor_position_changed = Signal(int, int)
     sig_new_file = Signal(str)
     sig_refresh_formatting = Signal(bool)
@@ -343,10 +210,6 @@ class CodeEditor(TextEditBaseWidget):
     # a URI is found, this signal is emitted
     sig_uri_found = Signal(str)
 
-    # Used for testing. When the mouse moves with Ctrl/Cmd pressed and
-    # the mouse left button is pressed, this signal is emitted
-    sig_go_to_uri = Signal(str)
-
     sig_file_uri_preprocessed = Signal(str)
     """
     This signal is emitted when the go to uri for a file has been
@@ -397,7 +260,7 @@ class CodeEditor(TextEditBaseWidget):
         self.current_project_path = None
 
         # Caret (text cursor)
-        self.setCursorWidth( CONF.get('main', 'cursor/width') )
+        self.setCursorWidth(CONF.get('main', 'cursor/width'))
 
         self.text_helper = TextHelper(self)
 
@@ -529,12 +392,10 @@ class CodeEditor(TextEditBaseWidget):
         # Indicate occurrences of the selected word
         self.cursorPositionChanged.connect(self.__cursor_position_changed)
         self.__find_first_pos = None
-        self.__find_flags = None
 
         self.language = None
         self.supported_language = False
         self.supported_cell_language = False
-        self.classfunc_match = None
         self.comment_string = None
         self._kill_ring = QtKillRing(self)
 
@@ -635,7 +496,6 @@ class CodeEditor(TextEditBaseWidget):
         self.hover_hints_enabled = None
 
         # Language Server
-        self.lsp_requests = {}
         self.filename = None
         self.completions_available = False
         self.text_version = 0
@@ -654,7 +514,6 @@ class CodeEditor(TextEditBaseWidget):
         self.range_formatting_enabled = False
         self.document_symbols_enabled = False
         self.formatting_characters = []
-        self.rename_support = False
         self.completion_args = None
         self.folding_supported = False
         self.is_cloned = False
@@ -670,7 +529,6 @@ class CodeEditor(TextEditBaseWidget):
         # Text diffs across versions
         self.differ = diff_match_patch()
         self.previous_text = ''
-        self.word_tokens = []
         self.patch = []
         self.leading_whitespaces = {}
 
@@ -1013,7 +871,8 @@ class CodeEditor(TextEditBaseWidget):
         # Indent guides
         self.toggle_identation_guides(indent_guides)
         if self.indent_chars == '\t':
-            self.indent_guides.set_indentation_width(self.tab_stop_width_spaces)
+            self.indent_guides.set_indentation_width(
+                self.tab_stop_width_spaces)
         else:
             self.indent_guides.set_indentation_width(len(self.indent_chars))
 
@@ -1034,7 +893,7 @@ class CodeEditor(TextEditBaseWidget):
 
         # Line number area and indent guides
         if cloned_from:
-            self.setFont(font) # this is required for line numbers area
+            self.setFont(font)  # this is required for line numbers area
             # Needed to show indent guides for splited editor panels
             # See spyder-ide/spyder#10900
             self.patch = cloned_from.patch
@@ -1096,7 +955,8 @@ class CodeEditor(TextEditBaseWidget):
 
         self.toggle_wrap_mode(wrap)
 
-        # Class/Function dropdown will be disabled if we're not in a Python file.
+        # Class/Function dropdown will be disabled if we're not in a Python
+        # file.
         self.classfuncdropdown.setVisible(show_class_func_dropdown
                                           and self.is_python_like())
 
@@ -1214,7 +1074,7 @@ class CodeEditor(TextEditBaseWidget):
         text = self.toPlainText()
         if self.is_ipython():
             # Send valid python text to LSP as it doesn't support IPython
-            text = ipython_to_python(text)
+            text = self.ipython_to_python(text)
         params = {
             'file': self.filename,
             'language': self.language,
@@ -1263,7 +1123,7 @@ class CodeEditor(TextEditBaseWidget):
         text = self.toPlainText()
         if self.is_ipython():
             # Send valid python text to LSP
-            text = ipython_to_python(text)
+            text = self.ipython_to_python(text)
         self.patch = self.differ.patch_make(self.previous_text, text)
         self.previous_text = text
         cursor = self.textCursor()
@@ -1367,7 +1227,7 @@ class CodeEditor(TextEditBaseWidget):
 
             def sort_key(completion):
                 if 'textEdit' in completion:
-                    text_insertion =  completion['textEdit']['newText']
+                    text_insertion = completion['textEdit']['newText']
                 else:
                     text_insertion = completion['insertText']
                 first_insert_letter = text_insertion[0]
@@ -1709,8 +1569,6 @@ class CodeEditor(TextEditBaseWidget):
         if edits is None:
             return
 
-        texts = []
-        diffs = []
         text = self.toPlainText()
         text_tokens = list(text)
         merged_text = None
@@ -1827,7 +1685,6 @@ class CodeEditor(TextEditBaseWidget):
     def update_and_merge_folding(self, extended_ranges):
         """Update and merge new folding information."""
         try:
-            text = self.toPlainText()
             folding_panel = self.panels.get(FoldingPanel)
 
             current_tree, root = merge_folding(
@@ -2034,7 +1891,7 @@ class CodeEditor(TextEditBaseWidget):
             for (key, value) in ALL_LANGUAGES.items():
                 if language.lower() in value:
                     self.supported_language = True
-                    sh_class, comment_string, CFMatch = self.LANGUAGES[key]
+                    sh_class, comment_string = self.LANGUAGES[key]
                     if key == 'IPython':
                         self.language = 'Python'
                     else:
@@ -2043,10 +1900,6 @@ class CodeEditor(TextEditBaseWidget):
                     if key in CELL_LANGUAGES:
                         self.supported_cell_language = True
                         self.has_cell_separators = True
-                    if CFMatch is None:
-                        self.classfunc_match = None
-                    else:
-                        self.classfunc_match = CFMatch()
                     break
 
         if filename is not None and not self.supported_language:
@@ -2093,6 +1946,7 @@ class CodeEditor(TextEditBaseWidget):
         """Get all cells."""
         if self.highlighter is None:
             return []
+
         # Filter out old cells
         def good(oedata):
             return oedata.is_valid() and oedata.def_type == oedata.CELL
@@ -2140,7 +1994,7 @@ class CodeEditor(TextEditBaseWidget):
             # blank import or dot completion
             self.do_completion()
         elif (leading_text.split()[0] in ['from', 'import'] and
-              not ';' in leading_text):
+                not ';' in leading_text):
             # import line with a single statement
             #  (prevents lines like: `import pdb; pdb.set_trace()`)
             self.do_completion()
@@ -2480,7 +2334,6 @@ class CodeEditor(TextEditBaseWidget):
         top = self.blockBoundingGeometry(block).translated(
                                                     self.contentOffset()).top()
         bottom = top + self.blockBoundingRect(block).height()
-        folding_panel = self.panels.get(FoldingPanel)
         while block.isValid() and top < event.pos().y():
             block = block.next()
             if block.isVisible():  # skip collapsed blocks
@@ -2603,7 +2456,6 @@ class CodeEditor(TextEditBaseWidget):
         """Overrides showEvent to update the viewport margins."""
         super(CodeEditor, self).showEvent(event)
         self.panels.refresh()
-
 
     #-----Misc.
     def _apply_highlighter_color_scheme(self):
@@ -2771,9 +2623,9 @@ class CodeEditor(TextEditBaseWidget):
             self.is_redoing = False
             self.skip_rstrip = False
 
-# =============================================================================
-#    High-level editor features
-# =============================================================================
+    # =========================================================================
+    #    High-level editor features
+    # =========================================================================
     @Slot()
     def center_cursor_on_next_focus(self):
         """QPlainTextEdit's "centerCursor" requires the widget to be visible"""
@@ -3126,7 +2978,6 @@ class CodeEditor(TextEditBaseWidget):
         """Get number of cells in document."""
         return 1 + len(list(self.cell_list()))
 
-
     #------Tasks management
     def go_to_next_todo(self):
         """Go to next todo and return new cursor position"""
@@ -3164,7 +3015,6 @@ class CodeEditor(TextEditBaseWidget):
             data.todo = message
             block.setUserData(data)
         self.sig_flags_changed.emit()
-
 
     #------Comments/Indentation
     def add_prefix(self, prefix):
@@ -3258,9 +3108,6 @@ class CodeEditor(TextEditBaseWidget):
                 cursor.movePosition(QTextCursor.EndOfBlock)
             cursor.endEditBlock()
         return number_spaces
-
-    def __is_cursor_at_start_of_block(self, cursor):
-        cursor.movePosition(QTextCursor.StartOfBlock)
 
     def remove_suffix(self, suffix):
         """
@@ -3875,6 +3722,7 @@ class CodeEditor(TextEditBaseWidget):
                 break
         if not __is_comment_bar(cursor1):
             return False
+
         def __in_block_comment(cursor):
             cs = self.comment_string
             return to_text_string(cursor.block().text()).startswith(cs)
@@ -3945,9 +3793,10 @@ class CodeEditor(TextEditBaseWidget):
         document = self.document()
         position -= 1
         while (position and not
-               is_letter_or_number(document.characterAt(position))):
+               self.is_letter_or_number(document.characterAt(position))):
             position -= 1
-        while position and is_letter_or_number(document.characterAt(position)):
+        while position and self.is_letter_or_number(
+                document.characterAt(position)):
             position -= 1
         cursor = self.textCursor()
         cursor.setPosition(position + 1)
@@ -3964,10 +3813,10 @@ class CodeEditor(TextEditBaseWidget):
         cursor.movePosition(QTextCursor.End)
         end = cursor.position()
         while (position < end and
-               not is_letter_or_number(document.characterAt(position))):
+               not self.is_letter_or_number(document.characterAt(position))):
             position += 1
         while (position < end and
-               is_letter_or_number(document.characterAt(position))):
+               self.is_letter_or_number(document.characterAt(position))):
             position += 1
         cursor.setPosition(position)
         return cursor
@@ -4839,7 +4688,6 @@ class CodeEditor(TextEditBaseWidget):
                       'upstream/origin remotes are not defined!'),
                     QMessageBox.Ok,
                 )
-        self.sig_go_to_uri.emit(uri)
         self.hide_tooltip()
         return full_uri
 
@@ -5002,7 +4850,6 @@ class CodeEditor(TextEditBaseWidget):
         self._last_point = pos
         alt = event.modifiers() & Qt.AltModifier
         ctrl = event.modifiers() & Qt.ControlModifier
-        shift = event.modifiers() & Qt.ShiftModifier
 
         if alt:
             self.sig_alt_mouse_moved.emit(event)
@@ -5193,7 +5040,6 @@ class CodeEditor(TextEditBaseWidget):
         top = int(self.blockBoundingGeometry(block).translated(
             self.contentOffset()).top())
         bottom = top + int(self.blockBoundingRect(block).height())
-        ebottom_top = 0
         ebottom_bottom = self.height()
 
         while block.isValid():
@@ -5283,37 +5129,38 @@ class CodeEditor(TextEditBaseWidget):
         """
         self.current_project_path = root_path
 
-#===============================================================================
-# CodeEditor's Printer
-#===============================================================================
+    def count_leading_empty_lines(self, cell):
+        """Count the number of leading empty cells."""
+        lines = cell.splitlines(keepends=True)
+        if not lines:
+            return 0
+        for i, line in enumerate(lines):
+            if line and not line.isspace():
+                return i
+        return len(lines)
 
-#TODO: Implement the header and footer support
-class Printer(QPrinter):
-    def __init__(self, mode=QPrinter.ScreenResolution, header_font=None):
-        QPrinter.__init__(self, mode)
-        self.setColorMode(QPrinter.Color)
-        self.setPageOrder(QPrinter.FirstPageFirst)
-        self.date = time.ctime()
-        if header_font is not None:
-            self.header_font = header_font
+    def ipython_to_python(self, code):
+        """Transform IPython code to python code."""
+        tm = TransformerManager()
+        number_empty_lines = self.count_leading_empty_lines(code)
+        try:
+            code = tm.transform_cell(code)
+        except SyntaxError:
+            return code
+        return '\n' * number_empty_lines + code
 
-    # <!> The following method is simply ignored by QPlainTextEdit
-    #     (this is a copy from QsciEditor's Printer)
-    def formatPage(self, painter, drawing, area, pagenr):
-        header = '%s - %s - Page %s' % (self.docName(), self.date, pagenr)
-        painter.save()
-        painter.setFont(self.header_font)
-        painter.setPen(QColor(Qt.black))
-        if drawing:
-            painter.drawText(area.right()-painter.fontMetrics().width(header),
-                             area.top()+painter.fontMetrics().ascent(), header)
-        area.setTop(area.top()+painter.fontMetrics().height()+5)
-        painter.restore()
+    def is_letter_or_number(self, char):
+        """
+        Returns whether the specified unicode character is a letter or a
+        number.
+        """
+        cat = category(char)
+        return cat.startswith('L') or cat.startswith('N')
 
 
-#===============================================================================
+# =============================================================================
 # Editor + Class browser test
-#===============================================================================
+# =============================================================================
 class TestWidget(QSplitter):
     def __init__(self, parent):
         QSplitter.__init__(self, parent)
